Re structure settings file names and paths
