+++ readthedocs/settings/dev/base.py
@@ -2,7 +2,7 @@
 from __future__ import absolute_import
 import os
 
-from .base import CommunityBaseSettings
+from ..base import CommunityBaseSettings
 
 
 class CommunityDevSettings(CommunityBaseSettings):
