+++ readthedocs/settings/dev/celery/webs.py
@@ -1,0 +1,4 @@
+"""Local setting for run celery web workers"""
+
+# Load same settings than ``readthedocs.settings.dev``
+from ..base import *
